requirements.txt
absl-py==2.1.0
ansitable==0.9.10
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
cffi==1.16.0
cfgv==3.4.0
colored==1.4.4
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
identify==2.5.33
install==1.3.5
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapipe==0.10.9
mpmath==1.3.0
nbformat==5.9.2
nest-asyncio==1.6.0
nodeenv==1.8.0
numpy==1.26.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pgraph-python==0.6.2
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
pre-commit==3.6.0
progress==1.6
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
roboticstoolbox-python==1.1.0
rpds-py==0.17.1
rtb-data==1.0.1
scipy==1.11.0
six==1.16.0
sounddevice==0.4.6
spatialgeometry==1.1.0
spatialmath-python==1.1.9
stack-data==0.6.3
swift-sim==1.1.0
sympy==1.12
tenacity==8.2.3
tk==0.1.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
virtualenv==20.25.0
wcwidth==0.2.13
websockets==12.0
